outreach_intel/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
agno>=1.0.0
anthropic>=0.40.0
